ISSUE #2865 auth route and processor renamed to users
